types fixed

+++ app/Models/UserSettings.php
@@ -38,9 +38,9 @@ class UserSettings extends Model
             'rain_enabled' => 'boolean',
             'snow_enabled' => 'boolean',
             'uvi_enabled' => 'boolean',
-            'rain_value' => 'integer',
-            'snow_value' => 'integer',
-            'uvi_value' => 'integer',
+            'rain_value' => 'float',
+            'snow_value' => 'float',
+            'uvi_value' => 'float',
         ];
     }
